requirements.txt
graphviz==0.20.1
Pillow==9.5.0
tk==0.1.0

